chore (redesign-ui): Pages in the process of being redesigned
- login & sign In pages
