+++ client/src/app/actions/fetchbooks.js
@@ -5,7 +5,7 @@ import {
   FETCH_BOOKS_BY_USER_ID,
   FETCH_BOOKS_FULFILLED_BY_USER_ID,
   FETCH_BOOKS_REJECTED_BY_USER_ID
-} from './type';
+} from './actiontype';
 import api from './api';
 
 export const fetchBooks = books => ({ type: FETCH_BOOKS, books });
