+++ client/src/app/components/presentation/common/Footer.jsx
