+++ client/src/app/components/presentation/common/header/NavigationBar.jsx
@@ -2,22 +2,24 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { Navbar, NavItem } from 'react-materialize';
 
-
+/**
+ * @description Component for Navigation
+ * @param {*} props
+ */
 const Navigation = (props) => {
   const links = props.navLinks.map(link =>
     (
       <li key={link}
         className={props.activeLink === link ? 'active' : ''}>
         <NavLink to={`/${link.replace(' ', '')}`}
-          activeClassName="active"
-        >
+          activeClassName="active">
           {link.toUpperCase()}
         </NavLink>
       </li>
     )
   );
   return (
-    <Navbar right id="nav-change" className='transparent' fixed={true}>
+    <Navbar right fixed={true}>
       {links}
     </Navbar>
   );
