+++ client/src/app/utils/setAuthorizationToken.js
@@ -10,15 +10,3 @@ const setAuthorizationToken = (token) => {
 
 export default setAuthorizationToken;
 
-
-// const setAuthorizationToken = (token) => {
-//   let headers = new Headers();
-//   if (token) {
-//     header['x-access-token'] = token;
-//     axios.defaults.headers.common['x-access-token'] = token;
-//   } else {
-//     delete axios.defaults.headers.common['x-access-token'];
-//   }
-// };
-
-// export default setAuthorizationToken;
