+++ server/src/controllers/userbooks.js
@@ -21,7 +21,7 @@ export default {
     if (!req.body.returndate) {
       return res.status(404).send({ message: 'Please specify a valid return date' });
     }
-    const returndate = (req.body.returndate).trim();
+    const returndate = req.body.returndate
     if (toDate(returndate) < Date.now() || !toDate(returndate)) {
       return res.status(422).send({ message: 'Please provide a valid return date' });
     }
@@ -76,9 +76,7 @@ export default {
             .catch(() => {
               res.status(404).send({ success: false, message: 'There is a problem with this user or book, Please contact the administrator' });
             });
-        }).catch((error) => {
-          res.status(400).send({ success: false, message: ` ${error.message}` });
-        });
+        })
       })
       .catch((error) => {
         res.status(400).send({ success: false, message: ` ${error.message}` });
@@ -198,9 +196,9 @@ export default {
                     });
                 }
               })
-              .catch(error => res.status(400).send(error.message));
+
           })
-          .catch(error => res.status(400).send(error.message));
+
       });
     }).catch(error => res.status(500).send(error.message));
   }
